Make sure we are in the correct group to run build-guest.
Problem:
build-guest fails for hard to determine reason.
Root cause, user was not running the command under the correct group.
They need to be in the mock group to run mock successfully.

Solution:
Verify that the command is being run under the correct group id.

Note: 'cgts' is our current alias for the 'mock' group.  Should probably
be renamed as stx-build in a future update.

Story: 2002835
Task: 24519
Change-Id: I658800cbae001f5cb0fc81bff1ed1ab159f6f6a3
Signed-off-by: Scott Little <scott.little@windriver.com>