Adding a build tool to create a helm charts tarball
The helm charts are packaged in rpms by the build process
and the purpose of this script is to package all those
artifacts together into a single tarball, for use by
other components in StarlingX.

Uses centos_helm.inc in each repo to determine what to search for.

Currently tested with packages from stx-config and stx-upstream:
 openstack-helm-infra
 openstack-helm
 stx-openstack-helm

Multiple tarballs are created, one per manifest.

Story: 2003909
Task: 27632
Depends-On: I3727a15fbdadc345a104c2af3446aee845171309
Change-Id: I06259d27e8f25d0b05eecdea49b2365246ac3112
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>