Change file permissions of RPM's to 644 in the ISO
This commit adds 644 permissions for all the files under
/www/pages/feed/<release>/Packages, allowing the web server to reach
all the required files to provision compute nodes.

The internal web server used to provide rpms for the installation of
computes and additional controller can fail to serve packages if these
has different permission less than the required by the web server.

lighttpd runs as a non-root user, however the content under
/www/pages/feed/<release>/Packages is owned by root. If the files hasn't
the "read for other" bit set, then this files won't be provided by the
web server and thus causing a failure in the installation of the
computes.

We ensure that all the RPM's has the 644 permissions in that
location after to build the StarlingX ISO image.

Closes-Bug: #1794894

Change-Id: I928b156dad1cffcda2d9a8d41631a56ed7ee29f7
Signed-off-by: Felipe de Jesus Ruiz Garcia <felipe.de.jesus.ruiz.garcia@intel.com>