Fix error handling for missing wheel package
A missing wheel RPM caused get-stx-wheels.sh to hang on a call
to rpm2cpio due to an empty string. This update improves the
error handling to fix the bash syntax on the file check to protect
against the empty string.

This update also corrects the usage of "local" keyword added to
the wheel and image scripts in a recent update, outside a function.

Change-Id: Ib8b83b376eacf44235a3d7fcd63d2451ca919210
Story: 2004751
Task: 29787
Signed-off-by: Don Penney <don.penney@windriver.com>