Update build-helm-charts.sh to include image versions
The build-stx-images.sh has been updated to store image
versions to files after building and pushing images.
This commit updates the chart build to allow specifying
the path to the image files/urls as input for building
image versions into armada manifest and the label of the
application tarball which is used to construct the tarball
name.
The default stable latest images in the manifest will be
used if there's no image file passed.

Change-Id: Ideea6ef3b9298a182be26fc1295456a978338f45
Story: 2005350
Task: 30419
Signed-off-by: Angie Wang <angie.wang@windriver.com>