Add pecan-1.3.3 to dev-wheels.cfg
The openstack/requirements master/upper-constraints.txt was
recently updated to point to pecan-1.3.3, but pypi.org does
not have a wheel for this newly released version. This, in
turn, caused a dev image build failure.

This update adds pecan-1.3.3 to the dev-wheels.cfg to build
a wheel for this version and add it to the dev wheels tarball.

Change-Id: I165a3b3e31dad275459d015936e4c9ea5a0e2d78
Closes-Bug: 1823984
Signed-off-by: Don Penney <don.penney@windriver.com>